+++ lib/src/plugin/pluto_pagination.dart
@@ -3,6 +3,8 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:pluto_grid/pluto_grid.dart';
 
+import '../ui/ui.dart';
+
 class PlutoPagination extends PlutoStatefulWidget {
   const PlutoPagination(
     this.stateManager, {
@@ -139,8 +141,7 @@ class PlutoPaginationState extends _PlutoPaginationStateWithChange {
 
   ButtonStyle _getNumberButtonStyle(bool isCurrentIndex) {
     return TextButton.styleFrom(
-      primary: Colors.transparent,
-      onSurface: Colors.transparent,
+      disabledForegroundColor: Colors.transparent,
       shadowColor: Colors.transparent,
       padding: const EdgeInsets.fromLTRB(5, 0, 0, 10),
       backgroundColor: Colors.transparent,